functions/static-site-deployer/requirements.txt
# Azure Functions for Python
azure-functions>=1.18.0
azure-functions-worker>=1.5.0

# Azure SDK
azure-storage-blob>=12.17.0
azure-identity>=1.13.0

# HTTP client for Static Web Apps API
aiohttp>=3.8.5
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
